Rewrite language detection and remove `LanguageDropdownContainer`
